READ ME: Overall Description of idea
This project is a Social Networking application designed to help people better their physique while adding a competitive aspect to the sport.  This application will be designed so that people of all sizes and physiques can compete with each other in weekly challenges. These challenges will be scored based on how much more you can lift each week.The scoring of these competitions will make it possible for a beginner to beat an experienced weightlifter.
This interface will not only be to help make it fun in the gym but will also help you reach your goals and keep you from falling behind the rest of the crowd when it comes to making gains. 

How it will work:
At the beginning of a new workout program you will measure your maximum repetitions of each lift required. You will then record your weight and then it will calculate percentages of your maximum weight to better help yourself push and improve each day. 

Point system:
There is a point system that will measure your set, reps, and the percentage of your max. The higher percentage of you max that you lift with, the more points you earn. 
Idea:
******
So 3x10 you’re supposed to do 65% of your max. I think we should do a completion bonus or 10 plus 1 extra point for every percent you do over the suggested 65 % and if you do extra reps that is an also an extra point for every extra rep.

Also, if youre doing set of 12 or 8 or anything that does not give you a percentage of your max that is not divisible by 5 it gives you leeway with what weight you want to do. You can either round up or round down to the nearest percentage that ends in a 0 or 5 giving you options. I believe this will help users who are not HARDCORE so they do not feel the sense of failure by not pushing hard, but it will not affect or skew the point system in any way.